+++ frontend/components/text-classifier/labeling-rules/RuleLabelsDefinition.vue
@@ -67,15 +67,14 @@
         >Show less</base-button
       >
     </div>
-    <div v-else class="empty-labels">
-      <p>This doesn't have any labels yet.</p>
-      <p>
-        To create new rules you need al least two labels. It's highly
-        recommended to also annotate some records with these labels. Go to the
-        annotation mode to
-        <a href="#" @click.prevent="changeToAnnotationViewMode"
-          >create the labels and annotate some records</a
-        >.
+    <div v-else>
+      <BaseFeedbackComponent
+        :feedbackInput="inputForFeedbackComponent"
+        @on-click="goToSettings"
+        class="feedback-area"
+      />
+      <p class="help-message">
+        {{ messageNotLabels }}
       </p>
     </div>
     <p class="rule__info" v-if="ruleInfo">{{ ruleInfo }}</p>
@@ -106,12 +105,23 @@ export default {
       type: Boolean,
       default: false,
     },
+    labelsFromORM: {
+      type: Array,
+      required: true,
+    },
   },
   data: () => {
     return {
       searchText: "",
       shownLabels: DatasetViewSettings.MAX_VISIBLE_LABELS,
       selectedLabelsVModel: [],
+      inputForFeedbackComponent: {
+        message: "Action needed: Create labels in the dataset settings",
+        buttonLabels: [{ label: "Create labels", value: "CREATE_LABELS" }],
+        feedbackType: "ERROR",
+      },
+      messageNotLabels:
+        "To create new rules, you need al least two labels. We highly recommended starting by annotating some records with these labels.",
     };
   },
   computed: {
@@ -166,7 +176,7 @@ export default {
       return this.dataset.query.text;
     },
     labels() {
-      return this.dataset.labels.map((l) => ({ class: l, selected: false }));
+      return this.labelsFromORM.map((l) => ({ class: l, selected: false }));
     },
     filteredLabels() {
       return this.labels.filter((label) =>
@@ -192,6 +202,12 @@ export default {
           }
         }
       });
+    },
+    datasetName() {
+      return this.$route.params.dataset;
+    },
+    datasetWorkspace() {
+      return this.$route.params.workspace;
     },
   },
   watch: {
@@ -239,6 +255,15 @@ export default {
         value: "annotate",
       });
     },
+    goToSettings() {
+      this.$router.push({
+        name: "datasets-workspace-dataset-settings",
+        params: {
+          dataset: this.datasetName,
+          workspace: this.datasetWorkspace,
+        },
+      });
+    },
   },
 };
 </script>
@@ -273,12 +298,9 @@ export default {
 .label-button {
   @extend %item;
 }
-.empty-labels {
-  a {
-    outline: none;
-    color: $primary-color;
-    text-decoration: none;
-  }
+.help-message {
+  color: $black-37;
+  max-width: 480px;
 }
 .label-button {
   margin: 5px;